+++ docs/ChangeLog.md
@@ -1,6 +1,9 @@
 # Change Log for Argument Completer Scripts
 
-### Register-VNAWSArgumentCompleter, v1.2.0
+### Register-VNActiveDirectoryArgumentCompleter, v1.0.0, Dec 2021
+- added completer for arguments for parameter `-Properties` for the cmdlets `Get-ADComputer`, `Get-ADGroup`, `Get-ADOrganizationalUnit`, `Get-ADUser`, for easy discovery/specifying of choice properties to return; based on [Matt McNabb](https://mattmcnabb.github.io/)'s example from a while ago
+
+### Register-VNAWSArgumentCompleter, v1.2.0, Dec 2021
 - added completer for arguments for parameter `-Service` for cmdlets `Get-AWSCmdletName`, `Get-AWSService` (feature request in [Issue #9](https://github.com/vNugglets/PowerShellArgumentCompleters/issues/9))
 - added completer for arguments for parameter `-ApiOperation` for cmdlet `Get-AWSCmdletName` (feature request in [Issue #9](https://github.com/vNugglets/PowerShellArgumentCompleters/issues/9))
 - added Tool Tip info for SSM Document and SSM Parameter completions (creation and last modified date, respectively)